requirements.txt
py-cord
APScheduler
python-dotenv
pyyaml
pyprobs
lxml
